+++ src/app/globals.css
@@ -58,6 +58,7 @@
   --breakpoint-lg: 1480px;
 
   --color-my-black: #0f0f0f;
+  --color-gray-600 : #9F9F9F;
   --color-gray-500: #161616;
   --color-gray-400: #5a5a5a;
   --color-gray-300: #a4a4a4;
