+++ src/components/landings/LandingHeader.jsx
@@ -1,0 +1,7 @@
+export default function LadingHeader() {
+    return(
+        <header>
+            LadingHeader
+        </header>
+    )
+}
