+++ src/components/landings/LandingNotificationFeatureSection.jsx
@@ -1,0 +1,7 @@
+export default function LandingNotificationFeatureSection() {
+    return (
+        <div>
+            ...
+        </div>
+    )
+}
