+++ src/components/landings/LandingRandomPointBoxSection.jsx
@@ -1,0 +1,5 @@
+export default function LandingRandomPointBoxSection() {
+    return(
+        <div>..</div>
+    )
+}
